+++ info/search.h
@@ -72,22 +72,22 @@ int string_in_line (char *string, char *line);
    with "find" are passed a SEARCH_BINDING, and return an absolute position
    marker of the item being searched for.  "Find" functions return a value
    of -1 if the item being looked for couldn't be found. */
-int skip_whitespace (char *string);
-int skip_non_whitespace (char *string);
-int skip_whitespace_and_newlines (char *string);
-int skip_node_separator (char *body);
+size_t skip_whitespace (char *string);
+size_t skip_non_whitespace (char *string);
+size_t skip_whitespace_and_newlines (char *string);
+size_t skip_node_separator (char *body);
 
 long find_node_separator (SEARCH_BINDING *binding);
 long find_file_section (SEARCH_BINDING *binding, char *label);
 long find_node_in_binding (char *nodename, SEARCH_BINDING *binding);
 
-regmatch_t match_by_index (MATCH_STATE *state, int index);
+regmatch_t match_by_index (MATCH_STATE *state, size_t index);
 enum search_result match_in_match_list (MATCH_STATE *state,
-                     long start, long end, int dir, int *match_index);
+                     long start, long end, int dir, size_t *match_index);
 
 void free_matches (MATCH_STATE *state);
 int matches_ready (MATCH_STATE *state);
-int at_end_of_matches (MATCH_STATE *state, int index);
+int at_end_of_matches (MATCH_STATE *state, size_t index);
 void decide_if_in_match (long off, int *in_match, MATCH_STATE *matches,
                          size_t *match_index);
